backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-dotenv==1.0.0
redis==5.0.1
numpy==1.24.3
opencv-python==4.8.1.78
pillow==10.1.0
requests==2.31.0

